cea_srv/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0 